test: move domain tests to the right namespace
